disabled create tranche, added tx history to portfolio, bug fixes, css updates
