+++ src/ui/tables/portfolio/index.ts
@@ -2,3 +2,4 @@ export * from './your-borrows';
 export * from './your-positions';
 export * from './your-supplies';
 export * from './your-rewards';
+export * from './your-transactions';
