+++ src/ui/tables/portfolio/your-transactions.tsx
@@ -1,0 +1,75 @@
+import React from 'react';
+import { useUserHistory } from '@/api';
+import { useTransactionsContext } from '@/store';
+import { DEFAULT_NETWORK, NETWORKS, truncate } from '@/utils';
+import { useAccount } from 'wagmi';
+import { getNetwork } from '@wagmi/core';
+import { AssetDisplay } from '@/ui/components';
+
+export const YourTransactionsTable = () => {
+    const { address } = useAccount();
+    const { transactions } = useTransactionsContext();
+    const { queryUserTxHistory } = useUserHistory(address);
+
+    const network = getNetwork()?.chain?.unsupported
+        ? DEFAULT_NETWORK
+        : getNetwork()?.chain?.name?.toLowerCase() || DEFAULT_NETWORK;
+
+    const renderStatus = (status: 'error' | 'pending' | 'complete') => {
+        switch (status) {
+            case 'complete': {
+                return <span className="text-green-400">Success</span>;
+            }
+            case 'pending': {
+                return <span className="text-yellow-400">Pending</span>;
+            }
+            default: {
+                return <span className="text-red-600">Error</span>;
+            }
+        }
+    };
+
+    return (
+        <>
+            {queryUserTxHistory?.data?.length || transactions?.length ? (
+                <div className="flex flex-col pt-6 pb-2 px-2 divide-y dark:divide-neutral-800">
+                    {transactions?.map((el, i) => (
+                        <div className="flex items-center justify-between" key={`transaction-${i}`}>
+                            <a
+                                href={`${NETWORKS[network].explorer}/tx/${el.text}`}
+                                className="py-1"
+                            >
+                                {truncate(el.text, 6)}
+                            </a>
+                            {renderStatus(el.status)}
+                        </div>
+                    ))}
+                    {queryUserTxHistory?.data?.map((el, i) => (
+                        <div
+                            className="flex items-center justify-between"
+                            key={`query-transaction-${i}`}
+                        >
+                            <a
+                                href={`${NETWORKS[network].explorer}/tx/${el.txHash}`}
+                                className="py-1"
+                            >
+                                {truncate(el.txHash, 6)}
+                            </a>
+                            <span>{el.type}</span>
+                            <span className="flex gap-1 items-center">
+                                {el.amount} <AssetDisplay name={el.asset} size={'sm'} noText />
+                            </span>
+                            <span>{el.datetime.toLocaleDateString()}</span>
+                        </div>
+                    ))}
+                </div>
+            ) : (
+                <div className="flex justify-center items-center pt-12 pb-4">
+                    <span className="text-neutral-700 dark:text-neutral-400">
+                        No Transaction History Available
+                    </span>
+                </div>
+            )}
+        </>
+    );
+};
